tHE REQUIRED FILE
I observed that this problem can be solved in two ways. One is the simple random number generation and observation, the other is a simultaneous two coin experiment. Consider an analogy:
A normal coin has a half chance of a head or a tail.
However, when two coins are tossed simultaneously, we see three unordered outcomes:
H&H
H&T/T&H
T&T
WE CAN SEE THAT THIS METHOD CAN ALSO BE USED BUT IS MUCH COMPLICATED (AS WE HANDLE TWO VARIABLES HERE).